requirements.txt
graspnetAPI
matplotlib
numpy
open3d
opencv-python
Pillow
pyserial
scipy
# for camera
# pyk4a
# pyrealsense2
